Validate configuration earlier in reconciliation, set conditions
