+++ operator/pom.xml
@@ -40,6 +40,11 @@
             <groupId>jakarta.validation</groupId>
             <artifactId>jakarta.validation-api</artifactId>
         </dependency>
+        <dependency>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-hibernate-validator</artifactId>
+        </dependency>
+
         <!-- Needed to generate OLM bundle. If you're not interested in this,
         you can remove this dependency -->
         <dependency>
